+++ src/main/java/com/linkage/LinkageApplication.java
@@ -7,10 +7,10 @@ import com.linkage.controller.FirebaseController;
 import com.linkage.controller.GoogleMapsController;
 import com.linkage.controller.MessageProviderController;
 import com.linkage.controller.SubscriptionController;
+import com.linkage.controller.VendorController;
 import com.linkage.controller.SmsController;
 import com.linkage.controller.WebhookController;
 import com.linkage.utility.AuthFilter;
-import com.linkage.utility.RedisClient;
 
 import io.dropwizard.core.Application;
 import io.dropwizard.core.setup.Environment;
@@ -34,8 +34,6 @@ public class LinkageApplication extends Application<LinkageConfiguration> {
         ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
         Validator validator = validatorFactory.getValidator();
 
-        RedisClient.start();
-
         AuthFilter authFilter = new AuthFilter(configuration.getxApiKey(), configuration.getAuthorizationKey());
 
         environment.servlets().addFilter("auth-filter", authFilter)
@@ -48,9 +46,9 @@ public class LinkageApplication extends Application<LinkageConfiguration> {
         DigitapController digitapController = new DigitapController(configuration, validator);
         SubscriptionController subscriptionController = new SubscriptionController(configuration, validator);
         SmsController smsController = new SmsController(configuration, validator);
-
         ErupeeController erupeeController = new ErupeeController(configuration, validator);
         GoogleMapsController googleMapsController=new GoogleMapsController(configuration, validator);
+        VendorController vendorController=new VendorController(configuration, validator);
         environment.jersey().register(befiscController);
         environment.jersey().register(webhookController);
         environment.jersey().register(firebaseController);
@@ -60,6 +58,7 @@ public class LinkageApplication extends Application<LinkageConfiguration> {
         environment.jersey().register(erupeeController);
         environment.jersey().register(googleMapsController);
         environment.jersey().register(smsController);
+        environment.jersey().register(vendorController);
 
     }
 }
