+++ src/main/java/com/linkage/client/BefiscService.java
@@ -66,31 +66,4 @@ public class BefiscService extends BaseServiceClient {
         }
     }
 
-
-    public ApiResponse<Object> sendAadharOtp(SendAadharOtp body) {
-
-        Map<String, String> reqBody = new HashMap<>();
-        reqBody.put("aadhaar", body.getAadharNumber());
-        // Encode the string in Base64
-        String authString = configuration.getBefiscAuthKey();
-        MultivaluedHashMap<String, Object> header = new MultivaluedHashMap<>();
-        header.putSingle("content-type", "application/json");
-        header.putSingle("authkey", authString);
-        
-        final String url = configuration.getBefiscSendAadharOtpUrl();
-        return this.networkCallExternalService(url, "POST", reqBody, header);
-
-    }
-
-    public ApiResponse<Object> verifyAadharOtp(Map<String,Object> body) {
-
-        // Encode the string in Base64
-        String authString = configuration.getBefiscAuthKey();
-        MultivaluedHashMap<String, Object> header = new MultivaluedHashMap<>();
-        header.putSingle("content-type", "application/json");
-        header.putSingle("authkey", authString);
-        final String url = configuration.getBefiscVerifyAadharOtpUrl();
-        return this.networkCallExternalService(url, "POST", body, header);
-
-    }
 }
