+++ src/main/java/com/linkage/client/DigitapService.java
@@ -2,15 +2,11 @@ package com.linkage.client;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
-import java.util.HashMap;
-import java.util.Map;
 
 import com.linkage.LinkageConfiguration;
 import com.linkage.api.ApiResponse;
 
 import com.linkage.core.validations.DigitapSchema.GetCreditBureau;
-import com.linkage.core.validations.DigitapSchema.SendAadharOtp;
-import com.linkage.core.validations.DigitapSchema.VerifyAadharOtp;
 import com.linkage.utility.Helper;
 
 import jakarta.ws.rs.core.MultivaluedHashMap;
@@ -51,43 +47,6 @@ public class DigitapService extends BaseServiceClient {
         header.putSingle("Authorization", "Basic " + encodedAuthString);
 
         final String url = configuration.getDigitapUrl() + "credit_analytics/request";
-
-        return this.networkCallExternalService(url, "POST", body, header);
-
-    }
-
-    public ApiResponse<Object> sendAadharOtp(SendAadharOtp body) {
-
-        Map<String, String> reqBody = new HashMap<>();
-        reqBody.put("uniqueId", body.getUniqueId());
-        reqBody.put("uid", body.getAadharNumber());
-
-        // Encode the string in Base64
-        String authString = configuration.getDigitapClientId() + ":" + configuration.getDigitapClientSecret();
-        String encodedAuthString = Base64.getEncoder().encodeToString(authString.getBytes(StandardCharsets.UTF_8));
-
-        MultivaluedHashMap<String, Object> header = new MultivaluedHashMap<>();
-        header.putSingle("content-type", "application/json");
-        header.putSingle("authorization", encodedAuthString);
-
-        final String url = configuration.getDigitapKycUrl() + "ent/v3/kyc/intiate-kyc-auto";
-
-        return this.networkCallExternalService(url, "POST", reqBody, header);
-
-    }
-
-    public ApiResponse<Object> verifyAadharOtp(Map<String,Object> body) {
-
-
-        // Encode the string in Base64
-        String authString = configuration.getDigitapClientId() + ":" + configuration.getDigitapClientSecret();
-        String encodedAuthString = Base64.getEncoder().encodeToString(authString.getBytes(StandardCharsets.UTF_8));
-
-        MultivaluedHashMap<String, Object> header = new MultivaluedHashMap<>();
-        header.putSingle("content-type", "application/json");
-        header.putSingle("authorization", encodedAuthString);
-
-        final String url = configuration.getDigitapKycUrl() + "ent/v3/kyc/submit-otp";
 
         return this.networkCallExternalService(url, "POST", body, header);
