+++ src/main/java/com/linkage/controller/BefiscController.java
@@ -5,19 +5,15 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import com.linkage.LinkageConfiguration;
 import com.linkage.api.ApiResponse;
 import com.linkage.client.BefiscService;
 import com.linkage.core.constants.HspKeywords;
-import com.linkage.core.validations.DigitapSchema.SendAadharOtp;
-import com.linkage.core.validations.DigitapSchema.VerifyAadharOtp;
 import com.linkage.core.validations.GetBankDetailsByAccSchema;
 import com.linkage.core.validations.GetVpaByMobileSchema;
 import com.linkage.core.validations.GetVpaDetailsSchema;
 import com.linkage.utility.Helper;
-import com.linkage.utility.RedisClient;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolation;
@@ -229,95 +225,5 @@ public class BefiscController extends BaseController {
         }
         return Map.of("valid_hsp", valid, "keyword", key);
     }
-
-    @SuppressWarnings("unchecked")
-    @POST
-    @Path("/sendAadharOtp")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response sendAadharOtp(@Context HttpServletRequest request, SendAadharOtp body) {
-
-        // Validate the request body
-        final Set<ConstraintViolation<SendAadharOtp>> violations = validator.validate(body);
-        if (!violations.isEmpty()) {
-            // Construct error message from violations
-            final String errorMessage = violations.stream()
-                    .map(ConstraintViolation::getMessage)
-                    .collect(Collectors.joining("; "));
-            return response(Response.Status.BAD_REQUEST, new ApiResponse<>(false, errorMessage, null));
-        }
-
-        // Call the service to send the OTP
-        final ApiResponse<Object> sendAadharOtpResponse = this.befiscService.sendAadharOtp(body);
-        if (!sendAadharOtpResponse.getStatus()) {
-            return response(Response.Status.INTERNAL_SERVER_ERROR, sendAadharOtpResponse);
-        }
-
-        // Cast the response data to the appropriate type
-        final Map<String, Object> sendAadharOtpData = (Map<String, Object>) sendAadharOtpResponse.getData();
-        if (sendAadharOtpData == null) {
-            return response(Response.Status.EXPECTATION_FAILED,
-            new ApiResponse<>(false, "Send OTP failed", null));
-        }
-        final Map<String, Object> sendAadharOtpResult = (Map<String, Object>) sendAadharOtpData.get("result");
-        if (sendAadharOtpResult == null) {
-            return response(Response.Status.EXPECTATION_FAILED,
-            new ApiResponse<>(false, "Send OTP failed", sendAadharOtpData));
-        }
-        // Return success response
-        return response(Response.Status.OK,
-                new ApiResponse<>(true, "Send OTP success", Map.of("access_key", sendAadharOtpResult.get("requestId"))));
-    }
-
-
-    
-    @SuppressWarnings("unchecked")
-    @POST
-    @Path("/verifyAadharOtp")
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response verifyAadharOtp(@Context HttpServletRequest request, VerifyAadharOtp body) {
-
-        // Validate the request body
-        final Set<ConstraintViolation<VerifyAadharOtp>> violations = validator.validate(body);
-        if (!violations.isEmpty()) {
-            // Construct error message from violations
-            String errorMessage = violations.stream()
-                    .map(ConstraintViolation::getMessage)
-                    .collect(Collectors.joining("; "));
-            return response(Response.Status.BAD_REQUEST, new ApiResponse<>(false, errorMessage, null));
-        }
-
-        final Map<String, Object> befiscRequest = new HashMap<>();
-        befiscRequest.put("aadhaar", body.getAadharNumber());
-        befiscRequest.put("referenceId", body.getAccessKey());
-        befiscRequest.put("otp", body.getOtp());
-
-        // Call the service to send the OTP
-        final ApiResponse<Object> befiscResponse = this.befiscService.verifyAadharOtp(befiscRequest);
-        if (!befiscResponse.getStatus()) {
-            return response(Response.Status.INTERNAL_SERVER_ERROR, befiscResponse);
-        }
-        // Cast the response data to the appropriate type
-        final Map<String, Object> befiscResponseData = (Map<String, Object>) befiscResponse.getData();
-        if (befiscResponseData == null) {
-             // Return success response
-            return response(Response.Status.EXPECTATION_FAILED,
-        new ApiResponse<>(false, "Aadhar verification failed", null));
-        }
-
-        final Map<String,Object> befiscResultData = (Map<String, Object> ) befiscResponseData;
-
-        Map<String, Object> clientRequiredData = new HashMap<>();
-        clientRequiredData.put("file", "result.xml_file");
-        clientRequiredData.put("address", "result.splitAddress");
-        clientRequiredData.put("image", "result.image");
-        clientRequiredData.put("zip", "result.pincode");
-        Map<String, Object> mappedValues = Helper.getMappedValuesFromMap(befiscResultData, clientRequiredData);
-        // Return success response
-        return response(Response.Status.OK,
-                new ApiResponse<>(true, "aadhar verification success", mappedValues));
-    }
-
 
 }
